requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic[email]
passlib[bcrypt]
python-jose[cryptography]